fix: service users

+++ src/modules/users/users.service.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
@@ -17,14 +17,26 @@ export class UsersService {
   }
 
   findOne(id: string) {
+    const findUser = this.userRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException(`User not found!`);
+    }
     return this.userRepository.findOne(id);
   }
 
   update(id: string, updateUserDto: UpdateUserDto) {
+    const findUser = this.userRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException(`User not found!`);
+    }
     return this.userRepository.update(id, updateUserDto);
   }
 
   remove(id: string) {
+    const findUser = this.userRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException(`User not found!`);
+    }
     return this.userRepository.delete(id);
   }
 }
